Refactor auditing classes: replace AuditableEntity with Trail, update related configurations and methods for improved clarity and consistency
