+++ src/Audit/Stores/AuditableStore.cs
@@ -24,8 +24,8 @@ public class AuditableStore<TContext, TKey, TUserType, TUserKey>(TContext contex
 {
    private readonly TContext _context = context ?? throw new ArgumentNullException(nameof(context));
 
-   private DbSet<AuditableEntity<TKey, TUserType, TUserKey>> AuditsSet
-      => _context.Set<AuditableEntity<TKey, TUserType, TUserKey>>();
+   private DbSet<Trail<TKey, TUserType, TUserKey>> AuditsSet
+      => _context.Set<Trail<TKey, TUserType, TUserKey>>();
 
    /// <summary>
    /// Provides a queryable collection of audit trails associated with the specified identity user and key types.
@@ -34,7 +34,7 @@ public class AuditableStore<TContext, TKey, TUserType, TUserKey>(TContext contex
    /// This property acts as an interface to the underlying entity set of audit trails,
    /// allowing for LINQ-based querying and manipulation of audit trail data.
    /// </remarks>
-   public IQueryable<AuditableEntity<TKey, TUserType, TUserKey>> Audits
+   public IQueryable<Trail<TKey, TUserType, TUserKey>> Audits
       => AuditsSet;
 
    /// <summary>Indicates whether changes to the context are automatically persisted to the database upon certain operations.</summary>
@@ -45,7 +45,7 @@ public class AuditableStore<TContext, TKey, TUserType, TUserKey>(TContext contex
    /// <param name="audit">The audit entity to be created.</param>
    /// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
    /// <returns>A result containing the created audit entity, indicating success or failure with potential error information.</returns>
-   public async Task<Result<AuditableEntity<TKey, TUserType, TUserKey>>> CreateAsync(AuditableEntity<TKey, TUserType, TUserKey> audit, CancellationToken cancellationToken)
+   public async Task<Result<Trail<TKey, TUserType, TUserKey>>> CreateAsync(Trail<TKey, TUserType, TUserKey> audit, CancellationToken cancellationToken)
    {
       _context.Add(audit);
       try
@@ -65,7 +65,7 @@ public class AuditableStore<TContext, TKey, TUserType, TUserKey>(TContext contex
    /// <param name="audit">The audit entity with updated information to be persisted.</param>
    /// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
    /// <returns>A result containing the updated audit entity, indicating success or failure with potential error details.</returns>
-   public async Task<Result<AuditableEntity<TKey, TUserType, TUserKey>>> UpdateAsync(AuditableEntity<TKey, TUserType, TUserKey> audit, CancellationToken cancellationToken)
+   public async Task<Result<Trail<TKey, TUserType, TUserKey>>> UpdateAsync(Trail<TKey, TUserType, TUserKey> audit, CancellationToken cancellationToken)
    {
       _context.Attach(audit);
       _context.Update(audit);
@@ -87,7 +87,7 @@ public class AuditableStore<TContext, TKey, TUserType, TUserKey>(TContext contex
    /// <param name="audit">The audit entity to be deleted.</param>
    /// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
    /// <returns>A result containing the deleted audit entity, indicating success or failure with potential error information.</returns>
-   public async Task<Result<AuditableEntity<TKey, TUserType, TUserKey>>> DeleteAsync(AuditableEntity<TKey, TUserType, TUserKey> audit, CancellationToken cancellationToken)
+   public async Task<Result<Trail<TKey, TUserType, TUserKey>>> DeleteAsync(Trail<TKey, TUserType, TUserKey> audit, CancellationToken cancellationToken)
    {
       _context.Remove(audit);
 
@@ -108,9 +108,9 @@ public class AuditableStore<TContext, TKey, TUserType, TUserKey>(TContext contex
    /// <param name="id">The unique identifier of the audit entry to retrieve.</param>
    /// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
    /// <returns>A result containing the found audit entry if it exists, or null if not found, along with success or failure information.</returns>
-   public async Task<Result<AuditableEntity<TKey, TUserType, TUserKey>?>> FindByIdAsync(TKey id, CancellationToken cancellationToken)
+   public async Task<Result<Trail<TKey, TUserType, TUserKey>?>> FindByIdAsync(TKey id, CancellationToken cancellationToken)
    {
-      var audit = await _context.Set<AuditableEntity<TKey, TUserType, TUserKey>>()
+      var audit = await _context.Set<Trail<TKey, TUserType, TUserKey>>()
                                 .FindAsync(id, cancellationToken);
 
       return Result.Success(audit);
@@ -121,9 +121,9 @@ public class AuditableStore<TContext, TKey, TUserType, TUserKey>(TContext contex
    /// <param name="userId">The unique identifier of the user associated with the audit entry.</param>
    /// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
    /// <returns>A result containing the found audit entity or null if not found, indicating success or failure with potential error information.</returns>
-   public async Task<Result<AuditableEntity<TKey, TUserType, TUserKey>?>> FindByIdAsync(TKey id, TUserKey userId, CancellationToken cancellationToken)
+   public async Task<Result<Trail<TKey, TUserType, TUserKey>?>> FindByIdAsync(TKey id, TUserKey userId, CancellationToken cancellationToken)
    {
-      var audit = await _context.Set<AuditableEntity<TKey, TUserType, TUserKey>>()
+      var audit = await _context.Set<Trail<TKey, TUserType, TUserKey>>()
                                 .FirstOrDefaultAsync(a => a.Id.Equals(id) && a.UserId!.Equals(userId), cancellationToken);
       return Result.Success(audit);
    }
@@ -132,9 +132,9 @@ public class AuditableStore<TContext, TKey, TUserType, TUserKey>(TContext contex
    /// <param name="userId">The unique identifier of the user to find the associated audit entry.</param>
    /// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
    /// <returns>A result containing the found audit entity or null, indicating the operation's outcome with potential error information.</returns>
-   public async Task<Result<AuditableEntity<TKey, TUserType, TUserKey>?>> FindByUserIdAsync(TUserKey userId, CancellationToken cancellationToken)
+   public async Task<Result<Trail<TKey, TUserType, TUserKey>?>> FindByUserIdAsync(TUserKey userId, CancellationToken cancellationToken)
    {
-      var audit = await _context.Set<AuditableEntity<TKey, TUserType, TUserKey>>()
+      var audit = await _context.Set<Trail<TKey, TUserType, TUserKey>>()
                                 .FirstOrDefaultAsync(a => a.UserId!.Equals(userId), cancellationToken);
       return Result.Success(audit);
    }
@@ -144,9 +144,9 @@ public class AuditableStore<TContext, TKey, TUserType, TUserKey>(TContext contex
    /// <param name="id">The unique identifier of the audit entry to find.</param>
    /// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
    /// <returns>A result containing the found audit entry if it exists, or null if not found, along with success or failure status and potential error information.</returns>
-   public async Task<Result<AuditableEntity<TKey, TUserType, TUserKey>?>> FindByUserIdAsync(TUserKey userId, TKey id, CancellationToken cancellationToken)
+   public async Task<Result<Trail<TKey, TUserType, TUserKey>?>> FindByUserIdAsync(TUserKey userId, TKey id, CancellationToken cancellationToken)
    {
-      var audit = await _context.Set<AuditableEntity<TKey, TUserType, TUserKey>>()
+      var audit = await _context.Set<Trail<TKey, TUserType, TUserKey>>()
                                 .FirstOrDefaultAsync(a => a.UserId!.Equals(userId) && a.Id.Equals(id), cancellationToken);
 
       return Result.Success(audit);
