+++ src/Audit/Stores/IAuditableStore.cs
@@ -17,43 +17,43 @@ public interface IAuditableStore<TKey, TUserType, TUserKey> : IDisposable
    /// <param name="audit">The audit entity to be created.</param>
    /// <param name="cancellationToken">A CancellationToken to observe while waiting for the task to complete.</param>
    /// <returns>A task that represents the asynchronous operation, containing the result of the created audit.</returns>
-   Task<Result<AuditableEntity<TKey, TUserType, TUserKey>>> CreateAsync(AuditableEntity<TKey, TUserType, TUserKey> audit, CancellationToken cancellationToken);
+   Task<Result<Trail<TKey, TUserType, TUserKey>>> CreateAsync(Trail<TKey, TUserType, TUserKey> audit, CancellationToken cancellationToken);
 
    /// <summary>Updates an existing audit entry asynchronously.</summary>
    /// <param name="auditTrail">The audit entity to be updated.</param>
    /// <param name="cancellationToken">A CancellationToken to observe while waiting for the task to complete.</param>
    /// <returns>A task that represents the asynchronous operation, containing the result of the updated audit.</returns>
-   Task<Result<AuditableEntity<TKey, TUserType, TUserKey>>> UpdateAsync(AuditableEntity<TKey, TUserType, TUserKey> auditTrail, CancellationToken cancellationToken);
+   Task<Result<Trail<TKey, TUserType, TUserKey>>> UpdateAsync(Trail<TKey, TUserType, TUserKey> auditTrail, CancellationToken cancellationToken);
 
    /// <summary>Deletes an existing audit entry asynchronously.</summary>
    /// <param name="auditTrail">The audit entity to be deleted.</param>
    /// <param name="cancellationToken">A CancellationToken to observe while waiting for the task to complete.</param>
    /// <returns>A task that represents the asynchronous operation, containing the result of the deletion operation.</returns>
-   Task<Result<AuditableEntity<TKey, TUserType, TUserKey>>> DeleteAsync(AuditableEntity<TKey, TUserType, TUserKey> auditTrail, CancellationToken cancellationToken);
+   Task<Result<Trail<TKey, TUserType, TUserKey>>> DeleteAsync(Trail<TKey, TUserType, TUserKey> auditTrail, CancellationToken cancellationToken);
 
    /// <summary>Finds an audit entry by its identifier asynchronously.</summary>
    /// <param name="id">The unique identifier of the audit entry to be retrieved.</param>
    /// <param name="cancellationToken">A CancellationToken to observe while waiting for the task to complete.</param>
    /// <returns>A task that represents the asynchronous operation, containing the result of the audit entry if it exists.</returns>
-   Task<Result<AuditableEntity<TKey, TUserType, TUserKey>?>> FindByIdAsync(TKey id, CancellationToken cancellationToken);
+   Task<Result<Trail<TKey, TUserType, TUserKey>?>> FindByIdAsync(TKey id, CancellationToken cancellationToken);
 
    /// <summary>Finds an audit entry by its identifier and associated user identifier asynchronously.</summary>
    /// <param name="id">The unique identifier of the audit entry to be retrieved.</param>
    /// <param name="userId">The unique identifier of the user associated with the audit entry.</param>
    /// <param name="cancellationToken">A CancellationToken to observe while waiting for the task to complete.</param>
    /// <returns>A task that represents the asynchronous operation, containing the result of the requested audit entry.</returns>
-   Task<Result<AuditableEntity<TKey, TUserType, TUserKey>?>> FindByIdAsync(TKey id, TUserKey userId, CancellationToken cancellationToken);
+   Task<Result<Trail<TKey, TUserType, TUserKey>?>> FindByIdAsync(TKey id, TUserKey userId, CancellationToken cancellationToken);
 
    /// <summary>Finds an audit entry by the specified user ID asynchronously.</summary>
    /// <param name="userId">The unique identifier of the user to search for.</param>
    /// <param name="cancellationToken">A CancellationToken to observe while waiting for the task to complete.</param>
    /// <returns>A task that represents the asynchronous operation, containing the result of the audit entry associated with the specified user ID.</returns>
-   Task<Result<AuditableEntity<TKey, TUserType, TUserKey>?>> FindByUserIdAsync(TUserKey userId, CancellationToken cancellationToken);
+   Task<Result<Trail<TKey, TUserType, TUserKey>?>> FindByUserIdAsync(TUserKey userId, CancellationToken cancellationToken);
 
    /// <summary>Finds an audit entry based on the user ID and audit ID asynchronously.</summary>
    /// <param name="userId">The ID of the user associated with the audit entry.</param>
    /// <param name="id">The unique identifier of the audit entry.</param>
    /// <param name="cancellationToken">A CancellationToken to observe while waiting for the task to complete.</param>
    /// <returns>A task representing the asynchronous operation, containing the result of the found audit entry.</returns>
-   Task<Result<AuditableEntity<TKey, TUserType, TUserKey>?>> FindByUserIdAsync(TUserKey userId, TKey id, CancellationToken cancellationToken);
+   Task<Result<Trail<TKey, TUserType, TUserKey>?>> FindByUserIdAsync(TUserKey userId, TKey id, CancellationToken cancellationToken);
 }
